backend/requirements.txt
django>=4.0,<5.0
djangorestframework>=3.16.0
django-cors-headers>=4.7.0
django-environ>=0.12.0
mysqlclient>=2.2.7
